Add in initialize option for an initialized AcidState
I found this useful for storing settings in AcidState that can be loaded on runtime. You can also pass this AcidState instance to Snaplets. For instance you can have one application datatype that stores users and pass the AcidState to a custom AuthBackend that reads from it (most likely by making it a class instance).

This also lets it be more flexible by adding in the ability for remote access.